Cleaned up the benchmarks
