+++ rust/otap-dataflow/Cargo.toml
@@ -98,12 +98,12 @@ strip = true
 #panic = "abort"
 
 [profile.bench]
-# inherits = "release"
+inherits = "release"
 # Potentially override lto for faster builds:
-lto = "thin"
+# lto = "thin"
 # Potentially override debug/strip for easier debugging:
-debug = true
-strip = "none"
+# debug = true
+# strip = "none"
 
 # cargo build --profile release-debug
 [profile.release-debug]
