+++ rust/otap-dataflow/crates/otap/src/encoder.rs
@@ -272,9 +272,7 @@ mod test {
     use prost::Message;
 
     fn _generate_logs_for_verify_all_columns() -> LogsData {
-        LogsData::new(
-            (0..2)
-                .map(|_| {
+        LogsData::new(vec![
                     ResourceLogs::build(
                         Resource::build(vec![KeyValue::new(
                             "resource_attr1",
@@ -309,39 +307,11 @@ mod test {
                                 .flags(LogRecordFlags::TraceFlagsMask)
                                 .body(AnyValue::new_string("log_body"))
                                 .finish(),
-                            LogRecord::build(2_000_000_000u64, SeverityNumber::Info, "event1")
-                                .observed_time_unix_nano(3_000_000_000u64)
-                                .trace_id(vec![0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3])
-                                .span_id(vec![0, 0, 0, 0, 1, 1, 1, 1])
-                                .severity_text("Info")
-                                .attributes(vec![KeyValue::new(
-                                    "log_attr1",
-                                    AnyValue::new_string("log_val_1"),
-                                )])
-                                .dropped_attributes_count(3u32)
-                                .flags(LogRecordFlags::TraceFlagsMask)
-                                .body(AnyValue::new_string("log_body"))
-                                .finish(),
-                            LogRecord::build(2_000_000_000u64, SeverityNumber::Info, "event1")
-                                .observed_time_unix_nano(3_000_000_000u64)
-                                .trace_id(vec![0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3])
-                                .span_id(vec![0, 0, 0, 0, 1, 1, 1, 1])
-                                .severity_text("Info")
-                                .attributes(vec![KeyValue::new(
-                                    "log_attr1",
-                                    AnyValue::new_string("log_val_1"),
-                                )])
-                                .dropped_attributes_count(3u32)
-                                .flags(LogRecordFlags::TraceFlagsMask)
-                                .body(AnyValue::new_string("log_body"))
-                                .finish(),
                         ])
                         .finish(),
                     ])
                     .finish()
-                })
-                .collect::<Vec<_>>(),
-        )
+        ])
     }
 
     fn _test_encode_logs_verify_all_columns_generic<T>(logs_data: T)
@@ -350,22 +320,7 @@ mod test {
     {
         // verify that every column for each record batch gets encoded as the correct type
         let result = encode_logs_otap_batch(&logs_data);
-        // TODO re-add this?
-        // assert!(result.is_ok());
         let otap_batch = result.unwrap();
-
-        // TODO remove all this
-        let logs_batch = otap_batch.get(ArrowPayloadType::Logs).unwrap();
-        arrow::util::pretty::print_batches(&[logs_batch.clone()]).unwrap();
-
-        let log_attrs = otap_batch.get(ArrowPayloadType::LogAttrs).unwrap();
-        arrow::util::pretty::print_batches(&[log_attrs.clone()]).unwrap();
-
-        let resource_attrs = otap_batch.get(ArrowPayloadType::ResourceAttrs).unwrap();
-        arrow::util::pretty::print_batches(&[resource_attrs.clone()]).unwrap();
-
-        let scope_attrs = otap_batch.get(ArrowPayloadType::ScopeAttrs).unwrap();
-        arrow::util::pretty::print_batches(&[scope_attrs.clone()]).unwrap();
 
         // check that the logs record batch is what we expect
         let expected_log_batch = RecordBatch::try_new(
@@ -795,8 +750,6 @@ mod test {
         // - using default values
 
         let result = encode_logs_otap_batch(logs_data);
-        // TODO add this back in?
-        // assert!(result.is_ok());
         let otap_batch = result.unwrap();
 
         let expected_logs_batch = RecordBatch::try_new(
@@ -856,6 +809,35 @@ mod test {
         logs_data.encode(&mut logs_data_bytes).unwrap();
         _test_encode_logs_verify_nullability_generic(&RawLogsData::new(&logs_data_bytes));
     }
+
+
+    fn _generate_logs_no_scopes_data() -> LogsData {
+        LogsData::new(vec![
+            ResourceLogs::build(Resource::new(vec![])).schema_url("https://schema.opentelemetry.io/resource_schema").scope_logs(vec![]).finish()
+        ])
+    }
+
+    fn _test_logs_no_scopes_generic<T>(logs_data: &T)
+    where
+        T: LogsDataView
+    {
+        
+    }
+
+    #[test]
+    fn test_logs_no_scopes_proto_struct() {
+        let logs_data = _generate_logs_no_scopes_data();
+        _test_logs_no_scopes_generic(&logs_data);
+    }
+
+    #[test]
+    fn test_logs_no_scopes_proto_bytes() {
+        let logs_data = _generate_logs_no_scopes_data();
+        let mut logs_data_bytes = vec![];
+        logs_data.encode(&mut logs_data_bytes).unwrap();
+        _test_logs_no_scopes_generic(&RawLogsData::new(&logs_data_bytes));
+    }
+
 
     fn _generate_log_body_all_field_types_data() -> LogsData {
         let log_bodies = vec![
