+++ rust/otap-dataflow/crates/pdata-views/Cargo.toml
@@ -13,3 +13,6 @@ otel-arrow-rust = { workspace = true }
 
 [lints]
 workspace = true
+
+[features]
+bench = []
